Added counting

+++ src/db_processor/ProgressBar.java
@@ -2,12 +2,12 @@ package db_processor;
 
 public class ProgressBar
 {
-	private int cur;
-	private int max;
+	private long cur;
+	private long max;
 	private int width;
 	private String unit;
 	
-	public ProgressBar(int max, int width, String unit)
+	public ProgressBar(long max, int width, String unit)
 	{
 		this.cur = 0;
 		this.max = max;
@@ -16,7 +16,7 @@ public class ProgressBar
 		update();
 	}
 	
-	public void set_cur(int cur)
+	public void set_cur(long cur)
 	{
 		if (cur < 0) {cur = 0;}
 		else if (cur > max) {cur = max;}
@@ -24,7 +24,7 @@ public class ProgressBar
 		this.cur = cur;
 		update();
 	}
-	public void set_max(int max)
+	public void set_max(long max)
 	{
 		if (max < 0) {max = 0;}
 		else if (max < cur) {max = cur;}
@@ -41,14 +41,14 @@ public class ProgressBar
 	{
 		if (max == 0)
 		{
-			System.out.print(Integer.toString(cur) + unit + "\r");
+			System.out.print(Long.toString(cur) + unit + "\r");
 		}
 		else
 		{
 			int perc = (int) Math.floor(cur * 100 / max);
 			int bar = (int) Math.floor(cur * width / max);
 	
-			System.out.print(Integer.toString(perc) + "% [" + repeat('=', bar) + repeat(' ', width - bar) + "] " + Integer.toString(cur) + unit + "\r");
+			System.out.print(Integer.toString(perc) + "% [" + repeat('=', bar) + repeat(' ', width - bar) + "] " + Long.toString(cur) + unit + "\r");
 		}
 	}
 	private StringBuilder repeat(char c, int n)
